Moved helper functions to koda.go
